+++ src/main/java/de/uni_jena/cs/fusion/abecto/converter/StringToQueryConverter.java
@@ -13,54 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.uni_jena.cs.fusion.abecto.datatype;
+package de.uni_jena.cs.fusion.abecto.converter;
 
-import org.apache.jena.datatypes.BaseDatatype;
 import org.apache.jena.datatypes.DatatypeFormatException;
-import org.apache.jena.graph.impl.LiteralLabel;
 import org.apache.jena.query.Query;
 import org.apache.jena.query.QueryException;
 import org.apache.jena.query.QueryFactory;
 import org.apache.jena.query.Syntax;
 
-import de.uni_jena.cs.fusion.abecto.vocabulary.AV;
+import com.fasterxml.jackson.databind.util.StdConverter;
 
-public class SparqlQueryType extends BaseDatatype {
-
-	public SparqlQueryType() {
-		super(AV.SparqlQuery.getURI());
-	}
+public class StringToQueryConverter extends StdConverter<String, Query> {
 
 	@Override
-	public String unparse(Object value) {
-		if (value instanceof Query) {
-			return ((Query) value).toString();
-		}
-		throw new IllegalArgumentException(String.format("Value is not a %s.", Query.class.getCanonicalName()));
+	public Query convert(String value) throws DatatypeFormatException {
+		return StringToQueryConverter.apply(value);
 	}
 
-	@Override
-	public Query parse(String lexicalForm) throws DatatypeFormatException {
+	public static Query apply(String value) throws DatatypeFormatException {
 		try {
-			return QueryFactory.create(lexicalForm, Syntax.syntaxSPARQL);
+			return QueryFactory.create(value, Syntax.syntaxSPARQL);
 		} catch (QueryException e) {
 			throw new DatatypeFormatException("Not a valid SPARQL query.", e);
 		}
 	}
-
-	@Override
-	public boolean isValidLiteral(LiteralLabel lit) {
-		return equals(lit.getDatatype()) && isValidValue(lit.getValue());
-	}
-
-	@Override
-	public boolean isValidValue(Object valueForm) {
-		return valueForm instanceof Query;
-	}
-
-	@Override
-	public Class<?> getJavaClass() {
-		return Query.class;
-	}
-
 }