+++ src/main/java/de/uni_jena/cs/fusion/abecto/vocabulary/AV.java
@@ -70,8 +70,6 @@ public class AV {
 	public static final Resource relativeCoverage = ResourceFactory.createResource(namespace + "relativeCoverage");
 	public static final Property relevantResource = ResourceFactory.createProperty(namespace, "relevantResource");
 	public static final Resource ResourceOmission = ResourceFactory.createResource(namespace + "ResourceOmission");
-	public static final Resource SparqlPropertyPath = ResourceFactory.createResource(namespace + "SparqlPropertyPath");
-	public static final Resource SparqlQuery = ResourceFactory.createResource(namespace + "SparqlQuery");
 	public static final Resource Step = ResourceFactory.createResource(namespace + "Step");
 	public static final Resource StepExecution = ResourceFactory.createResource(namespace + "StepExecution");
 	public static final Resource ValueOmission = ResourceFactory.createResource(namespace + "ValueOmission");